# restapi/swagger.json
{
    "openapi":"3.0.0",
    "info":{
        "description":"This is the api for user management",
        "version":"__version__"
    },
    "tags":[
        {
            "name":"PUBLIC ROUTE",
            "description": "PUBLIC ROUTE"
        }
    ],
    "paths":{
        "/health":{
            "summary":"Testing Health",
            "get":{
                "description": "Getting health of server",
                "tags":["PUBLIC ROUTE"],
                "responses":{
                    "200":{
                        "description": "Health OK"
                    }
                }
            }
        },
        "/users":{
            "summary":"Getting Users",
            "get":{
                "description": "Getting Users",
                "tags":["PUBLIC ROUTE"],
                "parameters":[
                    {
                        "in":"query",
                        "name":"role",
                        "schema":{
                            "type":"string"
                        },
                        "description": "Choose Role as Admin or User"
                    },
                    {
                        "in":"query",
                        "name":"city",
                        "schema":{
                            "type":"string"
                        },
                        "description": "Enter Your City"
                    }
                ],
                "responses":{
                    "200":{
                        "description": "Health OK"
                    }
                }
            }
        },
        "/user/{id}":{
            "summary":"Getting Particular Users",
            "get":{
                "description": "Getting Particular User",
                "tags":["PUBLIC ROUTE"],
                "parameters":[
                    {
                        "in":"path",
                        "name":"id",
                        "schema":{
                            "type":"string"
                        },
                        "description": "Choose Role as Admin or User"
                    }
                ],
                "responses":{
                    "200":{
                        "description": "Health OK"
                    }
                }
            }
        },
        "/addUser":{
            "summary":"Adding New User",
            "post":{
                "description": "Adding New User",
                "tags":["PUBLIC ROUTE"],
                "requestBody":{
                    "required":true,
                    "content":{
                        "application/json":{
                            "schema":{
                                "type":"object",
                                "id":"string"
                            },
                            "example":{
                                "name": "Amit",
                                "city": "Venice",
                                "phone": 345346457,
                                "role": "User",
                                "isActive": true
                            }
                        }
                    }
                },
                "responses":{
                    "200":{
                        "description": "Health OK"
                    }
                }
            }
        },
        "/updateUser":{
            "summary":"Update User",
            "put":{
                "description": "Update User",
                "tags":["PUBLIC ROUTE"],
                "requestBody":{
                    "required":true,
                    "content":{
                        "application/json":{
                            "schema":{
                                "type":"object",
                                "id":"string"
                            },
                            "example":{
                                "_id": "620b27b8406cc3c86a7be036",
                                "name": "Amit",
                                "city": "Venice",
                                "phone": 345346457,
                                "role": "User",
                                "isActive": true
                            }
                        }
                    }
                },
                "responses":{
                    "200":{
                        "description": "User Updated"
                    }
                }
            }
        },
        "/deleteUser":{
            "summary":"Delete User",
            "delete":{
                "description": "Delete User",
                "tags":["PUBLIC ROUTE"],
                "requestBody":{
                    "required":true,
                    "content":{
                        "application/json":{
                            "schema":{
                                "type":"object",
                                "id":"string"
                            },
                            "example":{
                                "_id": "620b27b8406cc3c86a7be036"
                            }
                        }
                    }
                },
                "responses":{
                    "200":{
                        "description": "User Updated"
                    }
                }
            }
        }
    }
}